move common.config tests into their own modules for better organization
